Move sorting tests into separate test file
